voxtral_4b_tts acoustic: the gate/up projections hand their [96, 9216] outputs to the SiLU-multiply in bf16 instead of fp32 (the multiply already emits bf16 for the bf8_b down projection); multiply 19.8->16.0 us per block, device 369.5->369.3 ms, trace acoustic 28.09->27.88 ms, full pipeline 373.2 ms (status ok), PCC 0.99961->0.999735

+++ models/tt_transformers/demo/voxtral_4b_tts_2603/_stubs/acoustic_transformer_block.py
@@ -507,9 +507,9 @@ def build(device, torch_module):
 
         hn = _block_norm(h4, eps, norm_scale, ttnn.bfloat16)
         gate = _lin(
-            hn, w1, dtype=ttnn.float32, compute_kernel_config=_TALL_COMPUTE, memory_config=ttnn.L1_MEMORY_CONFIG
+            hn, w1, dtype=ttnn.bfloat16, compute_kernel_config=_TALL_COMPUTE, memory_config=ttnn.L1_MEMORY_CONFIG
         )
-        up = _lin(hn, w3, dtype=ttnn.float32, compute_kernel_config=_TALL_COMPUTE, memory_config=ttnn.L1_MEMORY_CONFIG)
+        up = _lin(hn, w3, dtype=ttnn.bfloat16, compute_kernel_config=_TALL_COMPUTE, memory_config=ttnn.L1_MEMORY_CONFIG)
         h4 = ttnn.add(
             h4,
             _lin(
